requirements.txt
# Core dependencies
ollama>=0.1.0
termcolor>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# New dependencies for refactored version
pydantic>=2.0.0
dependency-injector>=4.0.0
structlog>=23.0.0
click>=8.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pyyaml>=6.0.0

# Optional dependencies
aiofiles>=23.0.0
asyncio-mqtt>=0.16.0 